+++ Att_multiplier.cpp
@@ -12,7 +12,7 @@ Att_multiplier& Att_multiplier::begin() {
   return *this;
 }
 
-Att_multiplier& Att_multiplier::onInput( atm_cb_t callback, int16_t idx ) {
+Att_multiplier& Att_multiplier::onInput( atm_cb_t callback, int idx ) {
   for ( uint8_t idx = 0; idx < ATM_MULTIPLIER_COMMS_MAX; idx++ ) {
     if ( _connector[idx].mode() == atm_connector::MODE_NULL ) {
       _connector[idx].set( callback, idx );
